refactor data collector
